style: renames, relocates
