--- nyx/moirai_ecs/system/nyx_system_base.py
@@ -1,8 +1,0 @@
-from abc import ABC
-
-from nyx.nyx_engine.managers.nyx_entity_manager import NyxEntityManager
-
-
-class NyxSystem(ABC):
-    def __init__(self, entity_manager: NyxEntityManager):
-        self.entity_manager: NyxEntityManager = entity_manager
