+++ nyx/moros_ecs/component/nyx_component.py
